+++ src/context/project.tsx
@@ -1,4 +1,5 @@
 import { captureException } from '@sentry/react'
+import { useAtomValue } from 'jotai'
 import {
   createContext,
   useCallback,
@@ -6,8 +7,9 @@ import {
   useEffect,
   useState,
 } from 'react'
-import { useLocation, useNavigate } from 'react-router-dom'
+import { useLocation, useNavigate, useParams } from 'react-router-dom'
 
+import { routeMatchForProjectPageAtom, useGetHistoryRoute } from '../config'
 import { getPath, PathName } from '../constants'
 import {
   useFundingFlow,
@@ -106,6 +108,11 @@ export const ProjectProvider = ({
   const [isProjectOwner, setIsProjectOwner] = useState<boolean | undefined>()
   const { user } = useAuthContext()
 
+  const params = useParams<{ projectId: string }>()
+  const routeMatchForProjectPage = useAtomValue(routeMatchForProjectPageAtom)
+  const historyRoutes = useGetHistoryRoute()
+  const lastRoute = historyRoutes[historyRoutes.length - 2] || ''
+
   const creatorModal = useModal()
   const milestonesModal = useModal()
   const rewardsModal = useModal<{
@@ -173,12 +180,23 @@ export const ProjectProvider = ({
       }
 
       if (project.id && project.owners[0]?.user.id === user.id) {
-        setIsProjectOwner(true)
+        if (
+          params.projectId &&
+          routeMatchForProjectPage &&
+          !lastRoute.includes('launch') &&
+          !(
+            lastRoute.includes('project') &&
+            lastRoute.includes(params.projectId)
+          )
+        ) {
+          navigate(getPath('projectOverview', `${params.projectId}`))
+          setIsProjectOwner(true)
+        }
       } else {
         setIsProjectOwner(false)
       }
     },
-    [user],
+    [user, params.projectId, routeMatchForProjectPage, lastRoute, navigate],
   )
 
   useEffect(() => {
