+++ src/pages/projectView/projectMainBody/ProjectMainBody.tsx
@@ -1,11 +1,4 @@
-import { useAtomValue } from 'jotai'
-import { Navigate, useParams } from 'react-router-dom'
-
-import {
-  routeMatchForProjectPageAtom,
-  useGetHistoryRoute,
-} from '../../../config'
-import { getPath } from '../../../constants'
+import Loader from '../../../components/ui/Loader'
 import { useProjectContext } from '../../../context'
 import { useProjectDetails } from '../projectNavigation/hooks/useProjectDetails'
 import {
@@ -23,19 +16,8 @@ export const ProjectMainBody = () => {
 
   const projectDetails = useProjectDetails(project)
 
-  const params = useParams<{ projectId: string }>()
-  const routeMatchForProjectPage = useAtomValue(routeMatchForProjectPageAtom)
-  const historyRoutes = useGetHistoryRoute()
-  const lastRoute = historyRoutes[historyRoutes.length - 2] || ''
-
-  if (
-    params.projectId &&
-    routeMatchForProjectPage &&
-    isProjectOwner &&
-    !lastRoute.includes('launch') &&
-    !(lastRoute.includes('project') && lastRoute.includes(params.projectId))
-  ) {
-    return <Navigate to={getPath('projectOverview', `${params.projectId}`)} />
+  if (isProjectOwner === undefined) {
+    return <Loader />
   }
 
   return (
